+++ frontend/src/lib/components/triggers/TriggersEditor.svelte
@@ -12,6 +12,7 @@
 	import type { TriggerContext } from '$lib/components/triggers'
 	import WebsocketTriggersPanel from './WebsocketTriggersPanel.svelte'
 	import ScheduledPollPanel from './ScheduledPollPanel.svelte'
+	import KafkaTriggersPanel from './KafkaTriggersPanel.svelte'
 
 	export let noEditor: boolean
 	export let newItem = false
@@ -31,6 +32,7 @@
 				<Tab value="schedules" selectedClass="text-primary text-sm font-semibold">Schedules</Tab>
 				<Tab value="routes" selectedClass="text-primary text-sm font-semibold">Routes</Tab>
 				<Tab value="websockets" selectedClass="text-primary text-sm font-semibold">Websockets</Tab>
+				<Tab value="kafka" selectedClass="text-primary text-sm font-semibold">Kafka</Tab>
 				<Tab value="emails" selectedClass="text-primary text-sm font-semibold">Email</Tab>
 				{#if isFlow}
 					<Tab value="scheduledPoll" selectedClass="text-primary text-sm font-semibold"
@@ -72,6 +74,12 @@
 					{#if $selectedTrigger === 'websockets'}
 						<div class="p-4">
 							<WebsocketTriggersPanel {newItem} path={currentPath} {isFlow} />
+						</div>
+					{/if}
+
+					{#if $selectedTrigger === 'kafka'}
+						<div class="p-4">
+							<KafkaTriggersPanel {newItem} path={currentPath} {isFlow} />
 						</div>
 					{/if}